implement site and affiliation filters. User exp is still wonky.
The only thing that actually filters which iframes show is clicking on a
filter list element. Thus, there's no way to clear filters.

Next tasks:
Users should be able to either delete the text in a filter or hit a
button to clear all text in a filter in order to remote that filter's
specificity.